Code review

+++ userhub/src/main/java/com/userhub/userhub/services/UserService.java
@@ -11,6 +11,8 @@ import com.userhub.userhub.core.entities.user.UserDto;
 import com.userhub.userhub.core.usecases.UserUsecase;
 import com.userhub.userhub.infrastructure.repositories.UserRepository;
 
+import jakarta.persistence.EntityNotFoundException;
+
 @Service
 public class UserService implements UserUsecase {
 
@@ -18,24 +20,32 @@ public class UserService implements UserUsecase {
     private UserRepository userRepository;
 
     public ResponseEntity<UserDto> saveUser(UserDto data) {
-        User newUser = new User(data.name(), data.cpf(),data.password(),data.email());
-        UserDto newUserDto = new UserDto(newUser.getName(), newUser.getCpf(), newUser.getPassword(),newUser.getEmail());
+
+        User newUser = new User(data.name(), data.cpf(), data.password(), data.email());
+        UserDto newUserDto = new UserDto(newUser.getName(), newUser.getCpf(), newUser.getPassword(),
+                newUser.getEmail());
         this.userRepository.save(newUser);
         return ResponseEntity.ok().body(newUserDto);
     }
 
-    public ResponseEntity<User> findUserById(String id) {
+    public ResponseEntity<UserDto> findUserById(String id) {
+        
         Optional<User> optinalUser = this.userRepository.findById(id);
-        if (optinalUser.isEmpty()) {
-            throw new NullPointerException("error teste");
-        }
-        User user = optinalUser.get();
-        return ResponseEntity.ok().body(user);
+        User user = optinalUser.orElseThrow(() -> new EntityNotFoundException(
+                "Nonexistent Record: The provided ID may not exist in the database, possibly due to deletion or an incorrect ID. "));
+                UserDto userDtoresponse = new UserDto(user.getName(),user.getCpf(), user.getEmail(),user.getPassword());
+                return ResponseEntity.ok().body(userDtoresponse);
+
     }
 
     public ResponseEntity<String> deleteUserById(String id) {
-        this.userRepository.deleteById(id);
-        return ResponseEntity.ok().body("Usuário deletado com Sucesso");
+        if (this.userRepository.existsById(id)) {
+            this.userRepository.deleteById(id);
+        } else {
+            throw new EntityNotFoundException(
+                    "Nonexistent Record: The provided ID may not exist in the database, possibly due to deletion or an incorrect ID. ");
+        }
+        return ResponseEntity.ok().body("User deleted successfully");
     }
 
 }
